Adding list utility class
This CL adds a simple list class.  In keeping with VapidSSL's constraint
to avoid any dynamic memory allocation, it requires the fixed size of
the list at creation time and backs the list with an array.  It is
conceptually similar to a std::vector<T> without resize.  Since C lacks
anything resmebling templates, the list functions are wrapped in macros
which take the element type.  This type is used to ensure sizeof(type)
matches the list's element size.  While this doesn't prevent type
confusion between two types of the same size, it should keep the list
structure sane.

Change-Id: Ia47d6274c27c5bb8d25ac79f001ca81f611a6af0