Add time to base/platform
This CL adds base/platform/time.h and the associated files.  On a POSIX
system, it just calls time(2).  Nonetheless, it's useful as a hook
point, both for other platforms which may get time differently, as well
as for testing, where base/platform/test/time_fake.h allows the
library's notion of time to manipulated directly.

Change-Id: Iefc4546ba22fc20a70299a3b7657f456ef9aad29