[Part 1/2] Update error API and test harness
This CL brings forward the public/error.h API, notably adding a few more
errors to the enums.  It also updates base/test/error_helper.h and
base/test/error_listener.h to more flexibly allow errors from other
sources, namely platforms and peers.

NOTE: The unit tests are NOT being compiled in this commit.  The main
unit test routine depends on the platform test harness, which in turn
depends on this CL.  That harness will be updated in a CL that
immediately follows this one, and that re-enables the unit tests.

Change-Id: I00a433af27825bc98ee9e704fca93325d4bdf7e8