[Part 2/2] Update platform test harness.
This CL updates the test harness in base/platform/test to integrate with
the immediately preceding updates to the error test harness, and to
better support multiple platforms.

NOTE: The previous CL needed to disable unit tests, as they depend on
these files which were not yet updated.  With this update, the unit
tests have been re-enabled.

Change-Id: Id40ac4895ea2f9d1347156d0097d282daba236e5