Update base/buf and base/test/scoped_buf
This CL adds some functionality to ScopedBuf, such as Raw() and Len(),
and reworks buf.c to make a few functions simpler.  In particular, it
eliminates buf_reserve in favor of an additional argument to buf_wrap,
and refactors buf_merge to correct a few bugs observed in later code.

Change-Id: I06e9fcdc6eba75d5589624f9637593c5958f6c84