Add AEAD algorithm adpater
This CL is from https://fuchsia-review.googlesource.com/#/c/3094.  It
includes only the AEAD specific portions used to encrypt and decrypt
data.

This CL includes AeadTest which is derived from CryptoTest, which uses
BoringSSLHelper to read test data files.  In the case of AEAD, the test
vector are from BoringSSL itself.

Change-Id: Ibb594b42386b0f396fe37263a05b32b508b13649