[Part 1/2] Add data streams
This CL adds common/stream.h, which arranges chunks into streams of
data.  It abstracts away some of the gnarly details of chunks, and
provides mechanisms for sending and receiving primitive data types and
buffers, for keeping track of nested data structures, and for keeping
rolling hashes of data being sent or received.

StreamHelper is incomplete here.  It was refactored as a result of
changes to CryptoHelper's code review.  Those changes are large enough
to warrant a separate CL.

Change-Id: Ia62659e8ec346fa0c4459a6010916e4aab5bd07b