Add platform-specific I/O routines.
This CL adds the bits necessary to send and receive network I/O. For
POSIX, this basically wraps send(2) and recv(2) without modifying the
attributes of the socket (i.e. it may be non-blocking, etc.)

This CL also includes io_mock.h, which allows unit tests to set up mock
connections, where the data is buffered.  It can be set up to loopback,
with data being sent then being available to be received. It can also be
set up to have two separate unidirectional connection buffers, one for
sending and one for receiving.

Change-Id: I19caf9201a11dea416e6b1cbf39104584e4c9224