[Part 2/2] Update StreamHelper
This CL refactors StreamHelper to allow it to use a test data file to
emulate a server.  While StreamHelper provides client streams as well,
it also has its methods arranged to allow subclasses to override those
methods and provide their own.

This refactor was prompted by changes to CryptoHelper, but ended up
pulling in substantial behavior that had previously been kept under
tls1_2/test.  By moving it here we are able to unify stream-based
testing, like that in x509v3.

Change-Id: Ifc08146a58046af7a4ed6bddae58aee99df5df81