Add X.509v3 certificate parsing and validation.
This CL adds both a certificate parser and a simple mechanism for
managing a list of trusted name/key pairs.  The certificate parser is
heavily styled after continuation passing in order to handle
non-blocking I/O with minimal buffering.

This parser does assume that certifcates follow RFC 5246, which
stipulates that certificates arrive in order starting from the leaf.
The root certificate may be omitted, as is common practice.

There is ample opportunity for more robust testing of almost every
failure case.  This is left as a TODO pending the project status.

Change-Id: I2351e1ce4ee8b10061b9b2d3ac19311eca5d0342