Add TLS 1.2 message layer
This CL builds on common/stream.h to create a TLS 1.2 message layer.
The code here differs in style from some of the other files with
continuation-passing style, such as x509v3/certificate.c.  This is
caused by the extra complexity of several more entry points.  It instead
uses a state variable to determine where to continue when a library
routine is retried after being interrupted by an I/O failure.

Change-Id: I36f867ea82b9cf9b556067a98da083f2a9db3976