Add TLS 1.2 PRF function
This CL implements the pseudo-random function defined by
https://tools.ietf.org/html/rfc5246#section-5.  It is essential that
this data be correct as it is used to generate keys for the bulk ciphers
and verify that the handshake was not tampered with.  As such, this CL
also includes a tool that links against libssl.a to generate test
vectors.

Note that this CL supports TLS 1.2 *only*; the change in definition of
the PRF was one of the more significant changes between TLS 1.2 and
earlier versions of the protocol.  In particular, the PRF will only use
the hash defined by the cipher suite; there is no MD5/SHA1 splitting.

Change-Id: I97272973e570d74a57c850e8b0560c658313cfc6