Add TLS 1.2 connection and related types
This CL adds tls1_2/tls.c, which is the fundamental class that
implements public/tls.h and represents a TLS connection.  The TLS
connection, handshake, and extensions are more tightly coupled than most
of the other compilation units thus far.  As a result this CL also
introduces the header files which define those objects as well.

Change-Id: Id41c92df4c6cad92a51cfeee617e1a3052229248